add method to communicate with services

+++ category-service/src/main/java/com/yas/categoryservice/controller/SalonCategoryController.java
@@ -4,6 +4,7 @@ package com.yas.categoryservice.controller;
 import com.yas.categoryservice.dto.SalonDTO;
 import com.yas.categoryservice.model.Category;
 import com.yas.categoryservice.service.CategoryService;
+import com.yas.categoryservice.service.client.SalonFeignClient;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -13,11 +14,12 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/categories/salon-owner")
 public class SalonCategoryController {
     private final CategoryService categoryService;
+    private final SalonFeignClient salonFeignClient;
 
     @PostMapping
-    public ResponseEntity<Category> createCategory(@RequestBody Category category) throws Exception {
-        SalonDTO salonDTO = new SalonDTO();
-        salonDTO.setId(1L);
+    public ResponseEntity<Category> createCategory(@RequestBody Category category,
+                                                   @RequestHeader("Authorization") String token) throws Exception {
+        SalonDTO salonDTO = salonFeignClient.getOwnerByOwnerId(token).getBody();
 
         Category savedCategory=categoryService.saveCategory(category,salonDTO);
         return ResponseEntity.ok(savedCategory);
@@ -25,10 +27,13 @@ public class SalonCategoryController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<String> deleteCategory(
-            @PathVariable("id") Long id
+            @PathVariable("id") Long id,
+            @RequestHeader("Authorization") String token
     ) throws Exception {
-        SalonDTO salonDTO = new SalonDTO();
-        salonDTO.setId(1L);
+        SalonDTO salonDTO = salonFeignClient.getOwnerByOwnerId(token).getBody();
+        if (salonDTO==null){
+            throw new Exception("User is not found");
+        }
 
         categoryService.deleteCategoryById(id, salonDTO.getId());
         return ResponseEntity.ok("Category deleted");
